Added Debug Levels to the MGE engine config. When set to ALL you receive the same logs as if no debug levels are enforced. Error/Debug Levels: NONE, //Nothing, NO errors FATAL_ERRORS_ONLY, //Atm not used directly.(when you only want to display errors(set to FATAL_ERRORS_ONLY) WARNINGS_ERRORS_ONLY, //Only output warnings and errors WARNINGS_ERRORS_LOG1, //Simple Logs that are indicating in not to much detail what the engine is doing WARNINGS_ERRORS_LOG2, //More Sophisticated Logs. Including stats of operation WARNINGS_ERRORS_LOG3, //Almost every log(except of logs in loops that get called a lot of times ALL //Everything

+++ Game/_vs2015/ParticleEmitter.cpp
@@ -159,7 +159,7 @@ void ParticleEmitter::UpdateParticles(float pTime)
 		_isEnabled = false;
 		return;
 	}
-	if (_activeParticles.size() < _maxParticles && !_stopProduce)
+	if ((int)_activeParticles.size() < _maxParticles && !_stopProduce)
 	{
 		SpawnParticles(glm::min(_maxParticles - totalActive, maxParticlesPerStep));
 		//std::cout << "Adding particles: " << std::to_string(_maxParticles - totalActive) << "\n";
@@ -185,7 +185,7 @@ void ParticleEmitter::UpdateParticles(float pTime)
 		}
 }
 
-void ParticleEmitter::SpawnParticles(int amount)
+void ParticleEmitter::SpawnParticles(size_t amount)
 {
 	for (size_t i = 0; i < amount; i++)
 	{
